+++ sources/types/str_find_char.c
@@ -1,4 +1,4 @@
-#include "polarium.h"
+#include "libfpp.h"
 
 char	*str_find_char(const char *s, int c, int jump)
 {
@@ -9,7 +9,7 @@ char	*str_find_char(const char *s, int c, int jump)
 	len = str_len(s);
 	while (s[++i] && s[i] != (char)c)
 		;
-	if (s[i] && s[i] == (char)c && jump < i - len)
+	if (s[i] && s[i] == (char)c && jump < i - (int)len)
 		return ((char *)&s[i + jump]);
 	return (NULL);
 }
